Minor improvements to make_release script.
Allow optional specification of a version string (the intended use of
this to be able to roll a tarball from a source tree / repo that's not
the exact version being released).

Compute and display SHA256 hash of the tarball on Linux, FreeBSD, and
MacOS.

Don't exclude make_release from the tarball (no point in hiding it)
but do try a little harder to get rid of .hg\* files.

Comment out some debugging code.